Revised README.
Although useful, the Lox language specification is not required reading for one to use the Lox interpreter. It has been moved to its own file.

The README file will henceforth focus on how to use the interpreter, along with the distinctions between the two main implementations (as described in Crafting Interpreters).